backend/app/services/web_scraper.py
import requests
from bs4 import BeautifulSoup
import time

class WebScraper:
    """Web scraping for plagiarism detection"""
    
    def __init__(self):
        self.headers = {
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36'
        }
        self.timeout = 10
        self.max_retries = 3
    
    def search_text(self, query, max_results=20):
        """Search for a query and return top results"""
        results = []
        
        try:
            # In production, use actual search API (Google, Bing, DuckDuckGo)
            # For demo, return empty list
            return results
        except Exception as e:
            print(f"Search error: {str(e)}")
            return results
    
    def _fetch_and_parse(self, url):
        """Fetch and parse webpage content"""
        try:
            response = requests.get(url, headers=self.headers, timeout=self.timeout)
            response.raise_for_status()
            
            soup = BeautifulSoup(response.content, 'html.parser')
            
            for script in soup(["script", "style"]):
                script.decompose()
            
            text = soup.get_text()
            lines = (line.strip() for line in text.splitlines())
            chunks = (phrase.strip() for line in lines for phrase in line.split("  "))
            text = ' '.join(chunk for chunk in chunks if chunk)
            
            return text[:10000]
        except Exception as e:
            print(f"Parse error: {str(e)}")
            return None